Updated NAS2D library -- Renderer::drawImageRepeated() now works as intended.
git-svn-id: https://svn.outpostuniverse.org:8443/svn/outposthd@236 bdf9137d-f3be-7946-88fe-7b0bbfdbf31a

+++ src/States/SplashState.cpp
@@ -112,7 +112,7 @@ State* SplashState::update()
 		r.drawImage(mLogoOutpostHd, r.center_x() - ((mLogoOutpostHd.width() + 250 * LOGO_SCALE)) / 2, r.center_y() - mLogoOutpostHd.height() / 2, LOGO_SCALE);
 
 
-	r.drawImage(mMousePointer, mMousePosition.x(), mMousePosition.y());
+	r.drawImage(mMousePointer, static_cast<float>(mMousePosition.x()), static_cast<float>(mMousePosition.y()));
 
 	if (r.isFading())
 		return this;
